requirements.txt
python-telegram-bot
fastapi
uvicorn
sqlite-utils
streamlit
python-dotenv
openai
supabase